pyproject.toml
[tool.poetry]
name = "medusa"
version = "0.1.0"
description = "Medusa: 4D face reconstruction and analysis"
authors = ["Lukas Snoek <lukassnoek@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://medusa.lukas-snoek.com/medusa"
repository = "https://github.com/medusa-4D/medusa"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
pandas = "^1.5.1"
numpy = "^1.23.4"
tqdm = "^4.64.1"
opencv-python = "^4.6.0.66"
scikit-image = "^0.19.3"
face-alignment = "^1.3.5"
PyYAML = "^6.0"
h5py = "^3.7.0"
trimesh = "^3.16.2"
pyrender = "^0.1.45"
torch = "^1.13.0"
torchvision = "^0.14.0"
mediapipe = "^0.8.11"
matplotlib = "^3.6.2"
onnxruntime = "^1.13.1"
gdown = "^4.5.3"

[tool.poetry.group.dev.dependencies]
jupyter-book = "^0.13.1"
sphinx-autoapi = "^2.0.0"
pytest = "^7.2.0"
nbval = "^0.9.6"

[tool.poetry.scripts]
medusa_videorecon = "medusa.cli:videorecon_cmd"
medusa_align = "medusa.cli:align_cmd"
medusa_resample = "medusa.cli:resample_cmd"
medusa_filter = "medusa.cli:filter_cmd"
medusa_preproc = "medusa.cli:preproc_cmd"
medusa_epoch = "medusa.cli:epoch_cmd"
medusa_videorender = "medusa.cli:videorender_cmd"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
